Add a post about income taxes.
This mostly just covers the marriage bonus / penalty and RSUs, nothing
super exciting or ground-breaking.

That said, I like to imagine that the rules of thumb I identify for "are
you going to see a marriage bonus or penalty" are broadly useful and
original, since I derived them from first principles.